+++ islands/HeaderCustom.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/headerCustom/HeaderCustom.tsx";
