+++ live.gen.ts
@@ -9,16 +9,17 @@ import * as $$$$1 from "./routes/styles.css.ts";
 import * as $$$$$0 from "./islands/OutOfStock.tsx";
 import * as $$$$$1 from "./islands/ProductList.tsx";
 import * as $$$$$2 from "./islands/HeaderSearchbar.tsx";
-import * as $$$$$3 from "./islands/HeaderModals.tsx";
-import * as $$$$$4 from "./islands/Newsletter.tsx";
-import * as $$$$$5 from "./islands/SearchControls.tsx";
-import * as $$$$$6 from "./islands/QuickViewModal.tsx";
-import * as $$$$$7 from "./islands/ProductImageZoom.tsx";
-import * as $$$$$8 from "./islands/SliderJS.tsx";
-import * as $$$$$9 from "./islands/WishlistButton.tsx";
-import * as $$$$$10 from "./islands/HeaderButton.tsx";
-import * as $$$$$11 from "./islands/ShippingSimulation.tsx";
-import * as $$$$$12 from "./islands/AddToCartButton.tsx";
+import * as $$$$$3 from "./islands/HeaderCustom.tsx";
+import * as $$$$$4 from "./islands/HeaderModals.tsx";
+import * as $$$$$5 from "./islands/Newsletter.tsx";
+import * as $$$$$6 from "./islands/SearchControls.tsx";
+import * as $$$$$7 from "./islands/QuickViewModal.tsx";
+import * as $$$$$8 from "./islands/ProductImageZoom.tsx";
+import * as $$$$$9 from "./islands/SliderJS.tsx";
+import * as $$$$$10 from "./islands/WishlistButton.tsx";
+import * as $$$$$11 from "./islands/HeaderButton.tsx";
+import * as $$$$$12 from "./islands/ShippingSimulation.tsx";
+import * as $$$$$13 from "./islands/AddToCartButton.tsx";
 import * as $$$$$$$$0 from "./sections/ProductShelf.tsx";
 import * as $$$$$$$$1 from "./sections/ProductList/ProductList.198.tsx";
 import * as $$$$$$$$2 from "./sections/DesignSystem.tsx";
@@ -27,9 +28,9 @@ import * as $$$$$$$$4 from "./sections/Features.tsx";
 import * as $$$$$$$$5 from "./sections/InteractiveBanner.tsx";
 import * as $$$$$$$$6 from "./sections/WishlistGallery.tsx";
 import * as $$$$$$$$7 from "./sections/CookieConsent.tsx";
-import * as $$$$$$$$8 from "./sections/Header2S/Header2S.tsx";
-import * as $$$$$$$$9 from "./sections/ProductDetails.tsx";
-import * as $$$$$$$$10 from "./sections/Footer.tsx";
+import * as $$$$$$$$8 from "./sections/ProductDetails.tsx";
+import * as $$$$$$$$9 from "./sections/Footer.tsx";
+import * as $$$$$$$$10 from "./sections/HeaderCustom/HeaderCustom.tsx";
 import * as $$$$$$$$11 from "./sections/Carousel.tsx";
 import * as $$$$$$$$12 from "./sections/SearchResult.tsx";
 import * as $$$$$$$$13 from "./sections/Header.tsx";
@@ -161,19 +162,20 @@ const manifest = {
     "./routes/styles.css.ts": $$$$1,
   },
   "islands": {
-    "./islands/AddToCartButton.tsx": $$$$$12,
-    "./islands/HeaderButton.tsx": $$$$$10,
-    "./islands/HeaderModals.tsx": $$$$$3,
+    "./islands/AddToCartButton.tsx": $$$$$13,
+    "./islands/HeaderButton.tsx": $$$$$11,
+    "./islands/HeaderCustom.tsx": $$$$$3,
+    "./islands/HeaderModals.tsx": $$$$$4,
     "./islands/HeaderSearchbar.tsx": $$$$$2,
-    "./islands/Newsletter.tsx": $$$$$4,
+    "./islands/Newsletter.tsx": $$$$$5,
     "./islands/OutOfStock.tsx": $$$$$0,
-    "./islands/ProductImageZoom.tsx": $$$$$7,
+    "./islands/ProductImageZoom.tsx": $$$$$8,
     "./islands/ProductList.tsx": $$$$$1,
-    "./islands/QuickViewModal.tsx": $$$$$6,
-    "./islands/SearchControls.tsx": $$$$$5,
-    "./islands/ShippingSimulation.tsx": $$$$$11,
-    "./islands/SliderJS.tsx": $$$$$8,
-    "./islands/WishlistButton.tsx": $$$$$9,
+    "./islands/QuickViewModal.tsx": $$$$$7,
+    "./islands/SearchControls.tsx": $$$$$6,
+    "./islands/ShippingSimulation.tsx": $$$$$12,
+    "./islands/SliderJS.tsx": $$$$$9,
+    "./islands/WishlistButton.tsx": $$$$$10,
   },
   "sections": {
     "$live/sections/PageInclude.tsx": i2$$$$0,
@@ -195,13 +197,13 @@ const manifest = {
     "deco-sites/time-198/sections/CookieConsent.tsx": $$$$$$$$7,
     "deco-sites/time-198/sections/DesignSystem.tsx": $$$$$$$$2,
     "deco-sites/time-198/sections/Features.tsx": $$$$$$$$4,
-    "deco-sites/time-198/sections/Footer.tsx": $$$$$$$$10,
+    "deco-sites/time-198/sections/Footer.tsx": $$$$$$$$9,
     "deco-sites/time-198/sections/Header.tsx": $$$$$$$$13,
-    "deco-sites/time-198/sections/Header2S/Header2S.tsx": $$$$$$$$8,
+    "deco-sites/time-198/sections/HeaderCustom/HeaderCustom.tsx": $$$$$$$$10,
     "deco-sites/time-198/sections/Highlights.tsx": $$$$$$$$15,
     "deco-sites/time-198/sections/InteractiveBanner.tsx": $$$$$$$$5,
     "deco-sites/time-198/sections/LinkTree.tsx": $$$$$$$$3,
-    "deco-sites/time-198/sections/ProductDetails.tsx": $$$$$$$$9,
+    "deco-sites/time-198/sections/ProductDetails.tsx": $$$$$$$$8,
     "deco-sites/time-198/sections/ProductList/ProductList.198.tsx": $$$$$$$$1,
     "deco-sites/time-198/sections/ProductShelf.tsx": $$$$$$$$0,
     "deco-sites/time-198/sections/QuickView/QuickView.198.tsx": $$$$$$$$14,
