--- sections/Header2S/Header2S.tsx
@@ -1,1 +1,0 @@
-export { default } from "$store/components/header2S/Header2S.tsx";
